Created Test class for constructing a colour object with a String with 6 Hexadecimals to represent the RGB values. Used an assert equals to check a known 6 digit hex value against its corresponding RGB values out of 255.
Wrote the corresponding methods in the colour class to create a colour object with a 6 hexadecimal string. Utilised a hashmap which each hex code in it. Split up the string into its 6 hex digits and convert them into a decimal number between 0 and 255 to represent the RGB colour.

create a method to get the 3 RGB values produced as a string for comparisons.

Implement an if statement that throws an Exception if hex string is not 6 digits long. and another if statement that checks that the provided string is a Hexadecimal.

All previous test cases pass.