Previous two commits created new methods getHex and getRGB, they also created two exceptions. These now need to be tested.
Created test to check exception is thrown when float number provided is over 1 and a test to check exception is thrown when float number provided is under 0.

Created a test to ensure exception is thrown when hex string provided is too long and a test to ensure exception is thrown when hex string provided is too short

Created a unit test to check that exception is thrown when Hex string provided contains a Non-Hex character.

All previous unit tests pass.